+++ CHANGELOG.md
@@ -8,6 +8,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Allow the do-block of `Spek.Macros.defcheck/2` to return an `:ok` or `:error`
+  tuple that references the check arguments.
+
 ## [0.4.0] - 2026-08-09
 
 ### Added